コンソール表示のフォーマットを変更
